+++ selene/support/jquery_style_selectors.py
@@ -19,6 +19,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+import warnings
 from typing import Union
 
 from selene.core.entity import Element, Collection
@@ -26,8 +27,14 @@ from selene.support.shared import browser
 
 
 def s(css_or_xpath_or_by: Union[str, tuple]) -> Element:
+    warnings.warn('selene.support.jquery_style_selectors.s is deprecated; '
+                  'use selene.support.shared.jquery_style.s instead',
+                  DeprecationWarning)
     return browser.element(css_or_xpath_or_by)
 
 
 def ss(css_or_xpath_or_by: Union[str, tuple]) -> Collection:
+    warnings.warn('selene.support.jquery_style_selectors.ss is deprecated; '
+                  'use selene.support.shared.jquery_style.ss instead',
+                  DeprecationWarning)
     return browser.all(css_or_xpath_or_by)