+++ selene/support/shared/jquery_style.py
@@ -19,7 +19,15 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+from typing import Union
 
-from selene.api.base import *
+from selene.core.entity import Element, Collection
+from selene.support.shared import browser
 
-from selene.support.shared import browser, config, s, ss
+
+def s(css_or_xpath_or_by: Union[str, tuple]) -> Element:
+    return browser.element(css_or_xpath_or_by)
+
+
+def ss(css_or_xpath_or_by: Union[str, tuple]) -> Collection:
+    return browser.all(css_or_xpath_or_by)